Update AaronSantiagoProvider.java
Switched to Short datatype for efficiency

+++ lesson_05/types/types_app/src/main/java/com/codedifferently/lesson5/dataprovider/AaronSantiagoProvider.java
@@ -13,7 +13,7 @@ public class AaronSantiagoProvider extends DataProvider {
   public Map<String, Class> getColumnTypeByName() {
     return Map.of(
         "column1", Long.class,
-        "column2", BigDecimal.class,
+        "column2", Short.class,
         "column3", Integer.class,
         "column4", Double.class,
         "column5", Boolean.class,
